@@ -4,74 +4,71 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
-    "start": "nx serve",
-    "start:admin": "nx serve admin",
+    "postinstall": "node ./decorate-angular-cli.js",
+    "dev": "nx serve",
     "build:prod": "nx build",
     "build:dev": "nx build --configuration development",
     "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "15.2.10",
-    "@angular/cdk": "15.2.9",
-    "@angular/common": "15.2.10",
-    "@angular/compiler": "15.2.10",
-    "@angular/core": "15.2.10",
-    "@angular/forms": "15.2.10",
-    "@angular/material": "15.2.9",
-    "@angular/platform-browser": "15.2.10",
-    "@angular/platform-browser-dynamic": "15.2.10",
-    "@angular/router": "15.2.10",
-    "@moby-it/poker-core": "^2.1.1",
-    "@ngneat/until-destroy": "^9.2.2",
-    "@ngrx/component": "15.3.0",
-    "@ngrx/component-store": "15.3.0",
-    "@ngrx/effects": "15.3.0",
-    "@ngrx/entity": "15.3.0",
-    "@ngrx/store": "15.3.0",
-    "@ngrx/store-devtools": "15.3.0",
-    "date-fns": "^2.29.3",
-    "ngx-json-viewer": "^3.2.1-alpha.0",
-    "ngx-toastr": "^15.2.1",
+    "@angular/animations": "17.1.3",
+    "@angular/cdk": "17.1.2",
+    "@angular/common": "17.1.3",
+    "@angular/compiler": "17.1.3",
+    "@angular/core": "17.1.3",
+    "@angular/forms": "17.1.3",
+    "@angular/material": "17.1.2",
+    "@angular/platform-browser": "17.1.3",
+    "@angular/platform-browser-dynamic": "17.1.3",
+    "@angular/router": "17.1.3",
+    "@moby-it/poker-core": "^2.1.3",
+    "@ngrx/component": "17.0.1",
+    "@ngrx/component-store": "17.0.1",
+    "@ngrx/effects": "17.0.1",
+    "@ngrx/entity": "17.0.1",
+    "@ngrx/store": "17.0.1",
+    "@ngrx/store-devtools": "17.0.1",
+    "date-fns": "^3.3.1",
+    "ngx-json-viewer": "^3.2.1",
+    "ngx-toastr": "^18.0.0",
     "rxjs": "7.8.1",
     "tslib": "^2.3.0",
-    "zone.js": "0.12.0",
-    "@nx/angular": "16.0.0"
+    "zone.js": "0.14.4",
+    "@nx/angular": "18.0.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "15.2.10",
-    "@angular-devkit/core": "15.2.10",
-    "@angular-devkit/schematics": "15.2.10",
-    "@angular-eslint/eslint-plugin": "15.0.0",
-    "@angular-eslint/eslint-plugin-template": "15.0.0",
-    "@angular-eslint/template-parser": "15.0.0",
-    "@angular/cli": "~15.2.0",
-    "@angular/compiler-cli": "15.2.10",
-    "@angular/language-service": "15.2.10",
-    "@nx/cypress": "16.0.0",
-    "@nx/eslint-plugin": "16.0.0",
-    "@nx/jest": "16.0.0",
-    "@nx/linter": "16.0.0",
-    "@nx/workspace": "16.0.0",
-    "@schematics/angular": "15.2.10",
+    "@angular-devkit/build-angular": "17.1.4",
+    "@angular-devkit/core": "17.1.4",
+    "@angular-devkit/schematics": "17.1.4",
+    "@angular-eslint/eslint-plugin": "17.0.1",
+    "@angular-eslint/eslint-plugin-template": "17.0.1",
+    "@angular-eslint/template-parser": "17.0.1",
+    "@angular/cli": "~17.1.0",
+    "@angular/compiler-cli": "17.1.3",
+    "@angular/language-service": "17.1.3",
+    "@nx/cypress": "18.0.4",
+    "@nx/eslint-plugin": "18.0.4",
+    "@nx/jest": "18.0.4",
+    "@nx/workspace": "18.0.4",
+    "@schematics/angular": "17.1.4",
     "@types/jest": "29.4.4",
-    "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "5.62.0",
-    "@typescript-eslint/parser": "5.62.0",
-    "cypress": "^12.2.0",
-    "eslint": "~8.15.0",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
+    "@types/node": "20.11.20",
+    "@typescript-eslint/eslint-plugin": "7.0.2",
+    "@typescript-eslint/parser": "6.21.0",
+    "cypress": "^13.0.0",
+    "eslint": "8.48.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-cypress": "2.15.1",
     "jest": "29.4.3",
     "jest-environment-jsdom": "29.4.3",
-    "jest-preset-angular": "13.0.0",
+    "jest-preset-angular": "14.0.3",
     "ng-mocks": "^14.2.3",
-    "nx": "16.0.0",
-    "nx-cloud": "18.0.0",
-    "prettier": "^2.6.2",
+    "nx": "18.0.4",
+    "prettier": "^3.2.5",
     "ts-jest": "29.1.2",
     "ts-node": "10.9.1",
-    "typescript": "4.9.5"
+    "typescript": "5.3.3",
+    "@nx/eslint": "18.0.4"
   }
 }
