@@ -1,16 +1,20 @@
 import { Component } from '@angular/core';
-import { NonNullableFormBuilder, Validators } from '@angular/forms';
+import { NonNullableFormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
 import { map, startWith } from 'rxjs';
 import { UserFormStore } from '../user-form.store';
+import { PushPipe } from '@ngrx/component';
+import { CommonModule } from '@angular/common';
 @Component({
   selector: 'ppo-signin-form',
   templateUrl: './signin-form.component.html',
+  imports: [PushPipe, CommonModule, ReactiveFormsModule],
+  standalone: true
 })
 export class SigninFormComponent {
   constructor(
     private componentStore: UserFormStore,
     private fb: NonNullableFormBuilder
-  ) {}
+  ) { }
   signinForm = this.fb.group({
     username: ['', [Validators.required]],
     password: ['', [Validators.required]],
