@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { LetModule } from '@ngrx/component';
+import { LetDirective } from '@ngrx/component';
 import { LeaderboardsPositionPipe } from './leaderboards-position.pipe';
 import { LeaderboardsComponent } from './leaderboards.component';
 import { LeaderboardsResolver } from './leaderboards.resolvers';
@@ -8,7 +8,7 @@ import { LeaderboardsRoutingModule } from './leaderboards.routing.module';
 import { LeaderboardsStore } from './leaderboards.store';
 
 @NgModule({
-  imports: [CommonModule, LeaderboardsRoutingModule, LetModule],
+  imports: [CommonModule, LeaderboardsRoutingModule, LetDirective],
   declarations: [LeaderboardsComponent, LeaderboardsPositionPipe],
   providers: [LeaderboardsResolver, LeaderboardsStore],
 })
