@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivate, Router } from '@angular/router';
+import { ActivatedRouteSnapshot, Router } from '@angular/router';
 import { PokerOddsFacade } from '@ppo/play/domain';
 import { map, Observable, skipWhile } from 'rxjs';
 @Injectable()
-export class ExistingGameGuard implements CanActivate {
+export class ExistingGameGuard  {
   constructor(private pokerFadace: PokerOddsFacade, private router: Router) {}
   canActivate(route: ActivatedRouteSnapshot): Observable<boolean> | boolean {
     const id = route.params['id'];
