@@ -1,1 +1,11 @@
-export { RoundUiModule } from './lib/round-ui.module';
+import { PokerTableComponent } from "./lib/poker-table/poker-table/poker-table.component";
+import { CopyRoundLinkButtonComponent } from "./lib/poker-table/round-actions/copy-link/copy-link.component";
+import { FavoriteButtonComponent } from "./lib/poker-table/round-actions/favorite/favorite.component";
+import { PlayRoundButtonComponent } from "./lib/poker-table/round-actions/play/play-round.component";
+import { RoundResultComponent } from "./lib/poker-table/round-result/round-result.component";
+export {
+  CopyRoundLinkButtonComponent, FavoriteButtonComponent,
+  PlayRoundButtonComponent, PokerTableComponent,
+  RoundResultComponent
+};
+
