@@ -4,5 +4,6 @@ import { Component } from '@angular/core';
   selector: 'ppo-loader',
   templateUrl: './loader.component.html',
   styleUrls: ['./loader.component.css'],
+  standalone: true
 })
-export class LoaderComponent {}
+export class LoaderComponent { }
