@@ -1,21 +1,22 @@
 import { CommonModule } from '@angular/common';
 import {
   Component,
+  DestroyRef,
   Input,
   OnChanges,
   OnInit,
   SimpleChanges,
+  inject,
 } from '@angular/core';
+import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
 import { ActivatedRoute, Router } from '@angular/router';
-import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { LetModule, PushModule } from '@ngrx/component';
+import { LetDirective, PushPipe } from '@ngrx/component';
 import { PokerOddsFacade } from '@ppo/play/domain';
 import { ScoreIsAccuratePipe } from '@ppo/round/domain';
-import { RoundUiModule } from '@ppo/round/ui';
+import { CopyRoundLinkButtonComponent, PlayRoundButtonComponent, PokerTableComponent, RoundResultComponent } from "@ppo/round/ui";
 import { UserProfileFacade, UserRoundViewmodel } from '@ppo/user/domain';
 import { BehaviorSubject, filter, switchMap, take, tap } from 'rxjs';
 import { SuitToSvgPipe } from './suitToSvg.pipe';
-@UntilDestroy()
 @Component({
   selector: 'ppo-round-list',
   templateUrl: './round-list.component.html',
@@ -29,12 +30,15 @@ import { SuitToSvgPipe } from './suitToSvg.pipe';
   ],
   standalone: true,
   imports: [
-    RoundUiModule,
     CommonModule,
-    PushModule,
-    LetModule,
+    PushPipe,
+    LetDirective,
     SuitToSvgPipe,
     ScoreIsAccuratePipe,
+    RoundResultComponent,
+    PlayRoundButtonComponent,
+    CopyRoundLinkButtonComponent,
+    PokerTableComponent
   ],
 })
 export class RoundListComponent implements OnChanges, OnInit {
@@ -43,10 +47,21 @@ export class RoundListComponent implements OnChanges, OnInit {
     private userProfile: UserProfileFacade,
     private route: ActivatedRoute,
     private router: Router
-  ) {}
+  ) {
+    this.route.queryParamMap.pipe(takeUntilDestroyed()).subscribe((params) => {
+      const selectedRoundId = params.get('selectedRoundId');
+      const selectedRoundIdx = this.rounds.findIndex(
+        (r) => r.roundId === selectedRoundId
+      );
+      if (selectedRoundIdx >= 0) {
+        this.selectedRound$.next(this.rounds[selectedRoundIdx]);
+      }
+    });
+  }
   selectedRound$ = new BehaviorSubject<UserRoundViewmodel | null>(null);
   watchingMyOwnmProfile$ = this.userProfile.watchingMyOwnProfile$;
   username$ = this.userProfile.username$;
+  destroy = inject(DestroyRef);
   @Input() rounds: UserRoundViewmodel[] = [];
   selectRound(round?: UserRoundViewmodel): void {
     if (!round) {
@@ -66,15 +81,7 @@ export class RoundListComponent implements OnChanges, OnInit {
         this.selectRound(this.rounds[0]);
       }
     }
-    this.route.queryParamMap.pipe(untilDestroyed(this)).subscribe((params) => {
-      const selectedRoundId = params.get('selectedRoundId');
-      const selectedRoundIdx = this.rounds.findIndex(
-        (r) => r.roundId === selectedRoundId
-      );
-      if (selectedRoundIdx >= 0) {
-        this.selectedRound$.next(this.rounds[selectedRoundIdx]);
-      }
-    });
+
   }
   ngOnChanges(changes: SimpleChanges): void {
     // if selected round not in new rounds, clear selected round
